+++ src/mulle-sde-migrate.sh
@@ -124,4 +124,13 @@ sde_migrate()
       oldmajor=0
       oldminor=42
    fi
+
+   #
+   # if craft etc is same as share now, we can remove etc
+   #
+   if diff -q .mulle/etc/craft .mulle/share/craft > /dev/null 2>&1
+   then
+      log_info "Removing .mulle/etc/craft as its no different from .mulle/share/craft"
+      rmdir_safer .mulle/etc/craft
+   fi
 }